requirements.txt
# Lambda-specific requirements
# These are the dependencies needed for Lambda functions
# Install with: pip install -r requirements-lambda.txt -t ./lambda-package

# Core web framework
fastapi==0.115.0
mangum==0.17.0
pydantic==2.9.2
email-validator>=2.1.0  # Required for Pydantic email validation
uvicorn[standard]>=0.30.0  # ASGI server for local development

# AWS and authentication
boto3>=1.34.0
authlib>=1.3.0
itsdangerous>=2.1.2
PyJWT>=2.8.0
cryptography>=42.0.5
passlib>=1.7.4
bcrypt>=4.0.1

# HTTP and utilities
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0

# Game engines
python-chess>=1.999

# Template rendering
jinja2>=3.1.4

# Security and rate limiting
slowapi>=0.1.9
python-multipart>=0.0.6





